requirements.txt
langchain 
langchain_core 
langchain_community 
langchain_groq 
langgraph
python-dotenv
streamlit
uvicorn